Methods for sim_result objects

Methods defined for objects returned from the ANOVA_exact, ANOVA_exact2, and ANOVA_power functions.

# S3 method for sim_result
print(x, ...)

# S3 method for sim_result
plot(x, ...)

# S3 method for sim_result
confint(object, parm = "main_results", level = 0.95, ...)

Arguments

x

object of class sim_result as returned from one of the simulation functions in Superpower.

...

further arguments passed through, see description of return value

object

Result returned from ANOVA_power (only applicable argument for confint)

parm

Argument for confint. Select what results from the simulation to return with confidence intervals. Options currently include: main_results (default), pc_results, manova_results, and emm_results.

level

Argument for confint. Confidence level for binomial proportion confidence intervals (Wilson, 1927). Default is .95.

Value

print

Prints short summary of the simulation result

plot

Returns meansplot or a plot of the distribution of p-values depending on whether an exact or Monte Carlo simulation was performed

confint

Returns confidence intervals for the selected result from ANOVA_power

References

Wilson, E. (1927). Probable Inference, the Law of Succession, and Statistical Inference. Journal of the American Statistical Association, 22(158), 209-212. doi:10.2307/2276774