Set/get global Superpower options

Global Superpower options are used, for example, by ANOVA_exact (et al.) and ANOVA_power. But can be changed in each functions directly using an argument (which has precedence over the global options).

Superpower_options(...)

Arguments

...

One of four: (1) nothing, then returns all options as a list; (2) a name of an option element, then returns its' value; (3) a name-value pair which sets the corresponding option to the new value (and returns nothing), (4) a list with option-value pairs which sets all the corresponding arguments. The example show all possible cases.

Value

depends on input, see above.

Details

The following arguments are currently set:

verbose should verbose (printed results) be set to true? Default is TRUE.

emm Option to perform analysis of estimated marginal means. Default is FALSE.

emm_model Model type ("multivariate", or "univariate") for estimated marginal means. Default is "multivariate".

contrast_type The type of comparison for the estimated marginal means. Default is "pairwise". See ?emmeans::'contrast-methods' for more details on acceptable methods.

plot Option to automatically print plots. Default is FALSE.

alpha_level Alpha level used to determine statistical significance. Default is .05.

correction Option to set a correction for sphericity violations. Default is no correction. This can be set to "none", "GG" Greenhouse-Geisser, and "HF" Huynh-Feldt

liberal_lambda Option to set a logical indicator of whether to use the liberal (cohen_f^2\*(num_df+den_df)) or conservative (cohen_f^2\*den_df) calculation of the noncentrality (lambda) parameter estimate. Default is FALSE.

Note

All options are saved in the global R options with prefix Superpower.